+++ packages/core/src/lib/fields/date.ts
@@ -1,7 +1,8 @@
 import { FieldTypes } from "@/enums/field-types.js";
 import { Field } from "@/lib/field.js";
+import { CoreError } from "../core-error.js";
 
-const toDate = (value: unknown) => {
+const toDate = (value: unknown): Date => {
   if (value instanceof Date) {
     return value;
   }
@@ -13,14 +14,17 @@ const toDate = (value: unknown) => {
     }
   }
 
-  return null;
+  throw new CoreError({
+    message: `Error serializing date with value ${value}`,
+  });
 };
 
 export class FieldDate extends Field<FieldTypes.DATE> {
   serializerMap: Field<FieldTypes.DATE>["serializerMap"] = {
     json: ({ value }) => {
       const date = toDate(value);
-      return date ? date.toJSON() : null;
+
+      return date.toJSON();
     },
     validation: ({ value }) => value,
     [Field.defaultSymbol]: ({ value }) => toDate(value),